changed to larger data types

+++ Mathematics/Fundamentals/Strange Grid Again.java
@@ -5,9 +5,9 @@ public class Solution {
 
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        int r = scan.nextInt();
+        long r = scan.nextLong();
         int c = scan.nextInt();
-        int num = (r%2 == 0) ? ((r-2)/2)*10 + 2*(c-1) + 1 : ((r-1)/2)*10 + 2*(c-1);
+        long num = (r%2 == 0) ? ((r-2)/2)*10 + 2*(c-1) + 1 : ((r-1)/2)*10 + 2*(c-1);
         System.out.println(num);
     }
 }
